# docs/api-specification.yml
openapi: 3.0.3
info:
  title: テストケース管理システム API
  description: |
    Next.js (App Router) + PostgreSQL + NextAuth.jsを使用したテストケース管理システムのREST API仕様書

    ## 認証
    すべてのAPIエンドポイント（認証関連を除く）はJWT認証が必要です。
    NextAuth.jsのセッショントークンを使用してリクエストを行います。

    ## 権限管理
    ### 静的ロール (user_role)
    - 0: 管理者 - すべての操作が可能
    - 1: テスト管理者 - テストグループの作成が可能
    - 2: 一般 - 割り当てられたテストグループのみアクセス可能

    ### 動的ロール (test_role)
    - 1: 設計者 - テストケースの作成・編集が可能
    - 2: 実施者 - テスト結果の登録が可能
    - 3: 閲覧者 - 閲覧のみ

  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: 開発環境
  - url: https://api.example.com
    description: 本番環境
...
paths:
...
  /test-groups:
    get:
      tags:
        - Test Groups
      summary: テストグループ一覧取得
      description: |
        ユーザーの権限に応じてアクセス可能なテストグループを取得
        - 管理者: 全テストグループ
        - テスト管理者: 自分が作成したグループ + 割り当てられたグループ
        - 一般: 割り当てられたグループのみ
      operationId: getTestGroups
      security:
        - bearerAuth: []
      responses:
        '200':
          description: テストグループ一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  testGroups:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestGroup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Test Groups
      summary: テストグループ作成
      description: |
        新しいテストグループを作成（管理者・テスト管理者のみ）
        **実装状態**: ✅ 完了
      operationId: createTestGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestGroupInput'
      responses:
        '201':
          description: テストグループ作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  testGroup:
                    $ref: '#/components/schemas/TestGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /test-groups/{groupId}:
    get:
      tags:
        - Test Groups
      summary: テストグループ詳細取得
      description: |
        指定したテストグループの詳細情報を取得
        **実装状態**: ✅ 完了
      operationId: getTestGroup
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: テストグループ詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  testGroup:
                    $ref: '#/components/schemas/TestGroup'
                  tags:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestGroupTag'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Test Groups
      summary: テストグループ更新
      description: |
        テストグループ情報を更新（作成者または管理者のみ）
        **実装状態**: ✅ 完了
      operationId: updateTestGroup
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestGroupInput'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  testGroup:
                    $ref: '#/components/schemas/TestGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Test Groups
      summary: テストグループ削除
      description: |
        テストグループをソフトデリート（作成者または管理者のみ）
        **実装状態**: ✅ 完了
      operationId: deleteTestGroup
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================
  # Test Cases
  # ==========================================
...
  /test-groups/{groupId}/report-data:
    get:
      tags:
        - Reports
      summary: 集計データ取得
      description: |
        テストグループの集計データを取得
        - 進捗率（実施済み/総件数）
        - OK/NG/未実施の件数
        - バグ曲線データ（日付ごとのNG件数）
        **実装状態**: 🔨 未実装（要実装）
      operationId: getReportData
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupId'
      responses:
        '200':
          description: 集計データ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==========================================
  # File Upload (S3)
  # ==========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth.js JWT トークン
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      description: テストグループID
      schema:
        type: integer
        example: 1
...
  schemas:
...
    TestGroup:
      type: object
      properties:
        id:
          type: integer
        oem:
          type: string
          example: Toyota
        model:
          type: string
          example: Camry
        event:
          type: string
          example: 量産準備
        variation:
          type: string
          example: グレードA
        destination:
          type: string
          example: 北米
        specs:
          type: string
          description: 仕様詳細
        test_startdate:
          type: string
          format: date
        test_enddate:
          type: string
          format: date
        ng_plan_count:
          type: integer
          description: 計画NG件数
        created_by:
          type: string
        updated_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TestGroupInput:
      type: object
      required:
        - oem
        - model
      properties:
        oem:
          type: string
        model:
          type: string
        event:
          type: string
        variation:
          type: string
        destination:
          type: string
        specs:
          type: string
        test_startdate:
          type: string
          format: date
        test_enddate:
          type: string
          format: date
        ng_plan_count:
          type: integer
        tags:
          type: array
          description: テストグループタグ
          items:
            type: object
            properties:
              tag_id:
                type: integer
              test_role:
                type: integer
                description: 1=設計者, 2=実施者, 3=閲覧者
    TestGroupTag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        test_role:
          type: integer

    # Test Cases
...
    ReportData:
      type: object
      properties:
        totalCases:
          type: integer
          description: 総テストケース数（is_target=true）
        completedCases:
          type: integer
          description: 実施済み件数
        okCases:
          type: integer
          description: OK件数
        ngCases:
          type: integer
          description: NG件数
        progressRate:
          type: number
          format: float
          description: 進捗率（%）
        bugCurve:
          type: array
          description: バグ曲線データ
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              ngCount:
                type: integer

    # Import
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: エラーメッセージ
  responses:
    BadRequest:
      description: リクエストが不正です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 権限がありません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []